Change Timeout Mechanism
Previously the program would stop if it didn't find a better solution after a certain number of iterations. Now it stops if it hasn't found a better solutino after a certain amount of time. TO-BE-MODIFIED comments were added to the "knobs" on our program. These "knobs" should help us to optomize.